requirements-dev.txt
# Development / Testing Dependencies
-r requirements.txt

pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pandas-stubs==2.2.3.241126
bandit==1.8.2
